+++ sdk/src/Core/AWSSDK.Core.NetFramework.csproj
@@ -70,4 +70,10 @@
     <Reference Include="System.Configuration" />
     <Reference Include="System.Net.Http" />
   </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="System.Buffers" Version="4.5.1" />
+    <PackageReference Include="System.Memory" Version="4.5.5" />
+    <PackageReference Include="System.Text.Json" Version="8.0.3" />
+  </ItemGroup>
 </Project>
