+++ sdk/src/Core/Amazon.Util/CryptoUtil.cs
@@ -17,9 +17,7 @@ using Amazon.Runtime.Internal.Util;
 using Amazon.Util.Internal;
 using AWSSDK.Runtime.Internal.Util;
 using System;
-#if NETCOREAPP3_1_OR_GREATER
 using System.Buffers;
-#endif
 using System.IO;
 using System.Security.Cryptography;
 using System.Text;
@@ -102,14 +100,13 @@ namespace Amazon.Util
             /// <returns>Computed hash code</returns>
             public string HMACSign(string data, string key, SigningAlgorithm algorithmName)
             {
-#if NETCOREAPP3_1_OR_GREATER
                 Encoding encoding = Encoding.UTF8;
                 int maxSize = encoding.GetMaxByteCount(data.Length);
                 byte[] buffer = ArrayPool<byte>.Shared.Rent(maxSize);
 
                 try
                 {
-                    int size = encoding.GetBytes(data, buffer);
+                    int size = encoding.GetBytes(data, 0, data.Length, buffer, 0);
                     ArraySegment<byte> binaryData = new ArraySegment<byte>(buffer, 0, size);
 
                     return HMACSign(binaryData, key, algorithmName);
@@ -118,10 +115,6 @@ namespace Amazon.Util
                 {
                     ArrayPool<byte>.Shared.Return(buffer);
                 }
-#else
-                byte[] binaryData = Encoding.UTF8.GetBytes(data);
-                return HMACSign(binaryData, key, algorithmName);
-#endif
             }
 
             /// <summary>
@@ -197,6 +190,95 @@ namespace Amazon.Util
             public string ComputeCRC32CHash(byte[] data)
             {
                 return ChecksumCRTWrapper.Crc32C(data);
+            }
+
+            public string HMACSign(byte[] data, string key, SigningAlgorithm algorithmName)
+                => HMACSign(new ArraySegment<byte>(data, 0, data.Length), key, algorithmName);
+
+            private string HMACSign(ArraySegment<byte> data, string key, SigningAlgorithm algorithmName)
+            {
+                if (String.IsNullOrEmpty(key))
+                    throw new ArgumentNullException("key", "Please specify a Secret Signing Key.");
+
+                if (data == null || data.Count == 0)
+                    throw new ArgumentNullException("data", "Please specify data to sign.");
+
+                KeyedHashAlgorithm algorithm = CreateKeyedHashAlgorithm(algorithmName);
+                if (null == algorithm)
+                    throw new InvalidOperationException("Please specify a KeyedHashAlgorithm to use.");
+
+                try
+                {
+                    algorithm.Key = Encoding.UTF8.GetBytes(key);
+                    byte[] bytes = algorithm.ComputeHash(data.Array, data.Offset, data.Count);
+                    return Convert.ToBase64String(bytes);
+                }
+                finally
+                {
+                    algorithm.Dispose();
+                }
+            }
+
+            public byte[] HMACSignBinary(byte[] data, byte[] key, SigningAlgorithm algorithmName)
+            {
+                if (key == null || key.Length == 0)
+                    throw new ArgumentNullException("key", "Please specify a Secret Signing Key.");
+
+                if (data == null || data.Length == 0)
+                    throw new ArgumentNullException("data", "Please specify data to sign.");
+
+                KeyedHashAlgorithm algorithm = CreateKeyedHashAlgorithm(algorithmName);
+                if (null == algorithm)
+                    throw new InvalidOperationException("Please specify a KeyedHashAlgorithm to use.");
+
+                try
+                {
+                    algorithm.Key = key;
+                    byte[] bytes = algorithm.ComputeHash(data);
+                    return bytes;
+                }
+                finally
+                {
+                    algorithm.Dispose();
+                }
+            }
+
+            static KeyedHashAlgorithm CreateKeyedHashAlgorithm(SigningAlgorithm algorithmName)
+            {
+                KeyedHashAlgorithm algorithm;
+                switch (algorithmName)
+                {
+                    case SigningAlgorithm.HmacSHA256:
+                        algorithm = new HMACSHA256();
+                        break;
+                    case SigningAlgorithm.HmacSHA1:
+                        algorithm = new HMACSHA1();
+                        break;
+                    default:
+                        throw new Exception(string.Format("KeyedHashAlgorithm {0} was not found.", algorithmName.ToString()));
+                }
+
+                return algorithm;
+            }
+
+
+            [ThreadStatic]
+            private static HashAlgorithm _hashAlgorithm = null;
+            private static HashAlgorithm SHA256HashAlgorithmInstance
+            {
+                get
+                {
+                    if (null == _hashAlgorithm)
+                    {
+                        _hashAlgorithm = CreateSHA256Instance();
+                    }
+                    return _hashAlgorithm;
+                }
+            }
+
+            internal static HashAlgorithm CreateSHA256Instance()
+            {
+                return SHA256.Create();
             }
         }
     }
